Updated Arrays.cls
Renamed the following members:
CreateInstance2 from CreateInstance4.  Now creates an instance of a two-dimensional array
CreateInstance3  from CreateInstance6  Now creates an instance of a three-dimensional array
CreateInstance4  now Creates a multidimensional Array of the specified Type and dimension lengths, with zero-based indexing. The dimension lengths are specified in an array of 32-bit integer
CreateInstance5 now Creates a multidimensional Array of the specified Type and dimension lengths, with zero-based indexing. The dimension lengths are specified in an array of 64-bit integer
CreateInstance6 now Creates a multidimensional Array of the specified Type and dimension lengths, with the specified lower bounds.

Updated affected examples
ArrayRankExample.bas